add level difficulty to actual game, edit routes, edit body color
