+++ src/client/src/App.js
@@ -2,10 +2,10 @@ import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Navbar from "./components/Navbar/Navbar";
 import LandingPage from "./pages/Landing Page/landing";
-import DailyPuzzle from "./pages/Daily/dailyPuzzle";
-import CustomizePage from "./pages/User Puzzle Design/customize";
+import DailyPuzzle from "./pages/Daily Puzzle/dailyPuzzle";
+import CustomizePage from "./pages/Design Puzzle/customize";
 import NotFound from "./pages/notFound";
-import Register from "./pages/Register/register";
+import Register from "./pages/Register Page/register";
 
 function App() {
 	return (
@@ -13,8 +13,8 @@ function App() {
 			<Navbar />
 			<Routes>
 				<Route exact path="/" element={<LandingPage />} />
-				<Route path="/daily" element={<DailyPuzzle />} />
-				<Route path="/customize" element={<CustomizePage />} />
+				<Route path="/game/dailypuzzle" element={<DailyPuzzle />} />
+				<Route path="/game/designpuzzle" element={<CustomizePage />} />
 				<Route path="/register" element={<Register />} />
 				<Route path="*" element={<NotFound />} />
 			</Routes>
