+++ src/client/src/components/Button/styles.css
@@ -1,7 +1,7 @@
 .defaultButton {
 	appearance: none;
-	background-color: #000000;
-	border: 2px solid #1a1a1a;
+	background-color: #df6c4f;
+	border: 2px solid #f28187;
 	box-sizing: border-box;
 	color: #ffffff;
 	cursor: pointer;
