+++ src/client/src/components/Navbar/styles.css
@@ -1,5 +1,5 @@
 .navbar {
-	background-color: purple;
+	background-color: #0c9d9e;
 	display: flex;
 	flex-wrap: wrap;
 	justify-content: space-between;
@@ -21,7 +21,7 @@
 	position: absolute;
 	top: 10px;
 	right: 30px;
-	}
+}
 .continueButton {
 	position: absolute;
 	top: 40px;
