+++ src/client/src/styles/index.css
@@ -5,7 +5,8 @@ body {
 		sans-serif;
 	-webkit-font-smoothing: antialiased;
 	-moz-osx-font-smoothing: grayscale;
-	background-color: grey;
+	background-color: #49c5b6;
+	color: white;
 }
 
 code {
